fix: move Terminal under Research; auto-run Backtest after Run analysis
- drop the top-nav Terminal entry; the view now lives as a tab in /research
- /console redirects to /research?tab=terminal so old links still resolve
- BacktestPanel auto-fires when a new analyze result lands (ref-guarded)
- VolRegime card notes that argmax_class needs a fine-tuned regime checkpoint
